+++ src/pages/MyPageHomePage/index.ts
@@ -1,0 +1,1 @@
+export { MyPageHomePage } from './MyPageHomePage';
