--- src/pages/MyPageJoinPage/index.ts
@@ -1,1 +1,0 @@
-export { EmailSignupPage } from './MyPageJoinPage';
